Add support for uploading custom user curves
In addition to the existing support for uploading interconnector price
curves, this adds a MultiCurveChooserView, which lists all curves that
may be customised by a visitor. The view allows the visitor to choose
from the available curves, perform an upload, and remove customised
curves.

Curve data is shown as returned by ETEngine (min, max, mean, etc).
Translations have been adjusted to make it easier to customise
descriptions based on the curve type, rather than unique to each curve
key. NL translations are pending.
