+++ app/assets/stylesheets/_accordion.sass
@@ -61,9 +61,7 @@
     margin: 0
     padding: 0
     list-style: none
-    overflow: hidden
     li.accordion_element
-      overflow: hidden
       vertical-align: top
       padding: 0 0 3px
       h3.selected a
