+++ app/views/scenarios/slides/_multi_curve_upload.html.haml
@@ -1,0 +1,2 @@
+.curve-upload.multi-curve-upload.custom-curve
+  .loading
