+++ app/views/scenarios/slides/_price_curve_upload.html.haml
@@ -1,6 +1,6 @@
-%p= t("custom_curves.#{curve_name}.description_html", docs_path: 'https://docs.energytransitionmodel.com/main/costs-imported-electricity#uploaded-price-curves')
+%p= t("custom_curves.types.#{curve_type}.description_html", docs_path: 'https://docs.energytransitionmodel.com/main/costs-imported-electricity#uploaded-price-curves')
 
-#import-price-curve-upload.curve-upload{ |
+#import-price-curve-upload.curve-upload.single-curve-upload{ |
   'data-curve-name' => curve_name, |
   'data-curve-disable-input' => associated_input, |
 }
