+++ config/interface/sidebar_items.yml
@@ -182,3 +182,7 @@
   section: supply
   tab_key: supply
   position: 9
+- key: curve_upload
+  section: flexibility
+  tab_key: flexibility
+  position: 10
