+++ config/interface/slides/flexibility_curve_upload.yml
@@ -1,0 +1,5 @@
+---
+- key: multi_curve_upload
+  position: 1
+  sidebar_item_key: curve_upload
+  output_element_key: workflow
