+++ config/locales/interface/en_sidebar_items.yml
@@ -122,6 +122,10 @@ en:
       title: Hydrogen costs
       short_description: ''
       description: ''
+    curve_upload:
+      short_title: Curve upload
+      title: Curve upload
+      description: Upload custom curves
     data_export:
       short_title: Data export
       title: Data export
