+++ config/locales/interface/nl_sidebar_items.yml
@@ -125,6 +125,10 @@ nl:
       title: Waterstofkosten
       short_description: ''
       description: ''
+    curve_upload:
+      short_title: Curve upload
+      title: Curve upload
+      description: Upload custom curves
     data_export:
       short_title: Data-export
       title: Data-export
