+++ config/locales/interface/slides/en_flexibility.yml
@@ -382,3 +382,7 @@ en:
         target =\"_blank\">documentation</a>.\r\n</p>\r\nDemand and production curves
         can be downloaded in the <a href=\"/scenario/data/data_export/merit-order-prices\">results
         section</a>."
+    multi_curve_upload:
+      title: Curve upload
+      description:
+        SLIDE DESCRIPTION
