+++ config/locales/interface/slides/nl_flexibility.yml
@@ -381,3 +381,7 @@ nl:
         target =\"_blank\">documentatie</a>.\r\n</p>\r\nProductiecurves kunnen worden
         gedownload in de <a href=\"/scenario/data/data_export/merit-order-prices\">resultaten
         sectie</a>.\r\n\r\n\r\n"
+    multi_curve_upload:
+      title: Curve upload
+      description:
+        SLIDE DESCRIPTION
