+++ config/locales/nl_custom_curves.yml
@@ -1,38 +1,49 @@
 ---
 nl:
   custom_curves:
-    confirm_remove: "Weet je zeker dat je dit profiel wilt verwijderen en de default wilt gebruiken?"
-    default: "Default profiel"
-    loading: "Laden..."
-    max: "Maximum"
-    mean: "Gemiddelde"
-    min: "Minimum"
+    confirm_remove: Weet je zeker dat je dit profiel wilt verwijderen en de default wilt gebruiken?
+    default: Default profiel
+    full_load_hours: Full load hours
+    loading: Laden...
+    max: Maximum
+    mean: Gemiddelde
+    min: Minimum
     on_date: "op %{date}"
-    remove: "Zet het default-profiel terug"
-    upload: "Een eigen profiel uploaden"
-    uploading: "Uploaden profiel"
+    remove: Zet het default-profiel terug
+    select_scenario: Selecteer scenario...
+    upload: Een eigen profiel uploaden
+    upload_from_scenario: Of gebruik een scenario
+    uploaded: geüpload
+    uploading: Uploaden profiel
+    use: Gebruik'
     value: "%{value}"
-    upload_from_scenario: "Of gebruik een scenario"
-    select_scenario: 'Selecteer scenario...'
-    use: 'Gebruik'
     errors:
-      file_too_large: "Profiel mag niet meer dan 1MB groot zijn"
-      illegal_value: "Profiel mag alleen numerieke waarden bevatten"
-      not_a_curve: "Profiel moet een CSV file zijn met 8760 getallen (één getal voor elk uur in een jaar)"
-      wrong_length: "Profiel moet 8760 getallen bevatten (één getal voor elk uur in een jaar)"
-    interconnector_1_price: &interconnector_price
-      default: "Default prijsprofiel"
-      default_description: "Constante prijs gedurende het jaar"
-      description_html: |
-        Het ETM veronderstelt standaard dat de prijs van geïmporteerde stroom constant is gedurende het jaar. Deze prijs kun je instellen met de slider hierboven. Je kunt afwijken van deze aanname door hieronder een
-        CSV-bestand te importeren met daarin voor elk uur van het jaar een prijs voor
-        import-stroom (in €/MWh). Voor meer uitleg over hoe dit werkt en voorbeelden van compatibele profielen
-        kun je <a href="%{docs_path}">onze documentatie bekijken.</a>.
-      value: "€%{value}/MWh"
-      curve_from_scenario: "Profiel vanuit scenario"
-      imported_from: 'Vanuit scenario'
-    interconnector_2_price: *interconnector_price
-    interconnector_3_price: *interconnector_price
-    interconnector_4_price: *interconnector_price
-    interconnector_5_price: *interconnector_price
-    interconnector_6_price: *interconnector_price
+      file_too_large: Profiel mag niet meer dan 1MB groot zijn
+      illegal_value: Profiel mag alleen numerieke waarden bevatten
+      not_a_curve: Profiel moet een CSV file zijn met 8760 getallen (één getal voor elk uur in een jaar)
+      wrong_length: Profiel moet 8760 getallen bevatten (één getal voor elk uur in een jaar)
+    names:
+      households_hot_water: Household hot water demand
+      interconnector_1_price: Interconnector 1 price
+      interconnector_2_price: Interconnector 2 price
+      interconnector_3_price: Interconnector 3 price
+      interconnector_4_price: Interconnector 4 price
+      interconnector_5_price: Interconnector 5 price
+      interconnector_6_price: Interconnector 6 price
+      weather/air_temperature: Air temperature
+    types:
+      price:
+        default: Default prijsprofiel
+        default_description: Constante prijs gedurende het jaar
+        description_html: |
+          Het ETM veronderstelt standaard dat de prijs van geïmporteerde stroom constant is gedurende het jaar. Deze prijs kun je instellen met de slider hierboven. Je kunt afwijken van deze aanname door hieronder een
+          CSV-bestand te importeren met daarin voor elk uur van het jaar een prijs voor
+          import-stroom (in €/MWh). Voor meer uitleg over hoe dit werkt en voorbeelden van compatibele profielen
+          kun je <a href=%{docs_path}>onze documentatie bekijken.</a>.
+        value: "€%{value}/MWh"
+        curve_from_scenario: Profiel vanuit scenario
+        imported_from: Vanuit scenario
+      temperature:
+        default: Default temperature curve
+        default_description: A temperature curve based on KNMI-data.
+        value: "%{value}°C"
