+++ global.d.ts
@@ -6,6 +6,9 @@ declare global {
   const App: {
     settings: {
       get: (key: string) => unknown;
+      set: (key: string, value: string | number | undefined) => unknown;
+      on: (eventName: string, event: (...args: unknown[]) => unknown) => unknown;
+      off: (eventName: string, event: (...args: unknown[]) => unknown) => unknown;
     };
   };
 }
